# cmd/api/docs/swagger.yaml
basePath: /
definitions:
...
  clap_internal_modules_league_dto.CreateLeagueRequest:
    properties:
      country:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      provider:
        type: string
      provider_league_id:
        type: string
    required:
    - name
    type: object
  clap_internal_modules_league_dto.UpdateLeagueRequest:
    properties:
      country:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      provider:
        type: string
      provider_league_id:
        type: string
    required:
    - name
    type: object
  clap_internal_modules_match_dto.CreateMatchRequest:
    properties:
      away_club_id:
        type: string
      home_club_id:
        type: string
      league_id:
        type: string
      match_datetime:
        type: string
      provider:
        type: string
      provider_match_id:
        type: string
      season_id:
        type: string
      stadium_name:
        type: string
      status:
        enum:
        - scheduled
        - live
        - halftime
        - finished
        - cancelled
        type: string
    required:
    - away_club_id
    - home_club_id
    - league_id
    - match_datetime
    - season_id
    - status
    type: object
  clap_internal_modules_match_dto.UpdateMatchRequest:
    properties:
      away_score:
        minimum: 0
        type: integer
      current_minute:
        maxLength: 10
        type: string
      home_score:
        minimum: 0
        type: integer
      match_datetime:
        type: string
      provider:
        type: string
      provider_match_id:
        type: string
      stadium_name:
        type: string
      status:
        enum:
        - scheduled
        - live
        - halftime
        - finished
        - cancelled
        type: string
    required:
    - match_datetime
    - status
    type: object
...
  clap_internal_shared_response.ErrorDetail:
    properties:
      code:
        type: integer
      details:
        items:
          type: string
        type: array
      message:
        type: string
    type: object
  clap_internal_shared_response.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/clap_internal_shared_response.ErrorDetail'
      message:
        type: string
      success:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    name: Clap API Support
  description: Clap match-day experience backend (auth, clubs, matches, realtime,
    shop).
  license:
    name: Proprietary
  termsOfService: http://swagger.io/terms/
  title: Clap Backend API
  version: "1.0"
paths:
...
  /api/v1/clubs/{id}/matches:
    get:
      parameters:
      - description: Club ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
      summary: List matches by club
      tags:
      - matches
...
  /api/v1/leagues:
    get:
      parameters:
      - description: Sort field
        in: query
        name: sort
        type: string
      - description: asc|desc
        in: query
        name: order
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
      summary: List leagues
      tags:
      - leagues
    post:
      consumes:
      - application/json
      parameters:
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/clap_internal_modules_league_dto.CreateLeagueRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Create league
      tags:
      - leagues
  /api/v1/leagues/{id}:
    delete:
      parameters:
      - description: League ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Delete league
      tags:
      - leagues
    get:
      parameters:
      - description: League ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
      summary: Get league
      tags:
      - leagues
    put:
      consumes:
      - application/json
      parameters:
      - description: League ID
        in: path
        name: id
        required: true
        type: string
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/clap_internal_modules_league_dto.UpdateLeagueRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Update league
      tags:
      - leagues
  /api/v1/leagues/{id}/matches:
    get:
      parameters:
      - description: League ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
      summary: List matches by league
      tags:
      - matches
...
  /api/v1/matches:
    get:
      parameters:
      - description: Sort field
        in: query
        name: sort
        type: string
      - description: asc|desc
        in: query
        name: order
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
      summary: List matches
      tags:
      - matches
    post:
      consumes:
      - application/json
      parameters:
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/clap_internal_modules_match_dto.CreateMatchRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Create match
      tags:
      - matches
  /api/v1/matches/{id}:
    delete:
      parameters:
      - description: Match ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Delete match
      tags:
      - matches
    get:
      parameters:
      - description: Match ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
      summary: Get match
      tags:
      - matches
    put:
      consumes:
      - application/json
      parameters:
      - description: Match ID
        in: path
        name: id
        required: true
        type: string
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/clap_internal_modules_match_dto.UpdateMatchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Update match
      tags:
      - matches
...
  /api/v1/matches/upcoming:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
      summary: List upcoming matches
      tags:
      - matches
...
  /api/v1/seasons/{id}/matches:
    get:
      parameters:
      - description: Season ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
      summary: List matches by season
      tags:
      - matches
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'JWT access token. Example: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"